+++ packages/eslint-config-airbnb-base/legacy.js
@@ -27,5 +27,7 @@ module.exports = {
       property: '__defineSetter__',
       message: 'Please use Object.defineProperty instead.',
     }],
+    'no-var': 'off',
+    strict: ['error', 'safe'],
   }
 };