+++ packages/eslint-config-airbnb/hooks.js
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: [
+    './rules/react-hooks.js',
+  ].map(require.resolve),
+  rules: {}
+};