+++ packages/eslint-config-airbnb/test/requires.js
@@ -1,7 +1,4 @@
-/* eslint
-   strict: 0,
-   global-require: 0,
- */
+/* eslint strict: 0, global-require: 0 */
 
 'use strict';
 
@@ -12,6 +9,7 @@ test('all entry points parse', (t) => {
   t.doesNotThrow(() => require('../base'), 'base does not throw');
   t.doesNotThrow(() => require('../legacy'), 'legacy does not throw');
   t.doesNotThrow(() => require('../whitespace'), 'whitespace does not throw');
+  t.doesNotThrow(() => require('../hooks'), 'hooks does not throw');
 
   t.end();
 });