+++ packages/eslint-config-airbnb-base/CHANGELOG.md
@@ -1,3 +1,42 @@
+14.0.0 / 2019-08-09
+==================
+ - [breaking] `no-self-assign`: enable `props` option
+ - [breaking] enable `no-useless-catch`
+ - [breaking] enable `max-classes-per-file`
+ - [breaking] enable `no-misleading-character-class`
+ - [breaking] enable `no-async-promise-executor`
+ - [breaking] enable `prefer-object-spread`
+ - [breaking] `func-name-matching`: enable `considerPropertyDescriptor` option
+ - [breaking] `padded-blocks`: enable `allowSingleLineBlocks` option (#1255)
+ - [breaking] `no-multiple-empty-lines`: Restrict empty lines at beginning of file (#2042)
+ - [breaking] Set 'strict' to 'never' (#1962)
+ - [breaking] legacy: Enable 'strict' (#1962)
+ - [breaking] Simplifies `no-mixed-operators` (#1864)
+ - [breaking] Require parens for arrow function args (#1863)
+ - [breaking] add eslint v6, drop eslint v4
+ - [patch] `camelcase`: enable ignoreDestructuring
+ - [patch] Add markers to spaced-comment block for Flow types (#1966)
+ - [patch] Do not prefer destructuring for object assignment expression (#1583)
+ - [deps] update `confusing-browser-globals`, `eslint-plugin-import`, `tape`, `babel-preset-airbnb`
+ - [dev deps] update babel-related deps to latest
+ - [dev deps] update `eslint-find-rules`, `eslint-plugin-import`
+ - [tests] only run tests in non-lint per-package travis job
+ - [tests] use `eclint` instead of `editorconfig-tools`
+
+13.2.0 / 2019-07-01
+==================
+ - [minor] Enforce dangling underscores in method names (#1907)
+ - [fix] disable `no-var` in legacy entry point
+ - [patch] Ignore property modifications of `staticContext` params (#2029)
+ - [patch] `no-extraneous-dependencies`: Add jest.setup.js to devDeps (#1998)
+ - [meta] add disabled `prefer-named-capture-group` rule
+ - [meta] add disabled `no-useless-catch` config
+ - [deps] Switch to confusing-browser-globals (#1961)
+ - [deps] update `object.entries`, `eslint-plugin-import`, `eslint-plugin-jsx-a11y`, `eslint-plugin-react`, `tape`
+ - [docs] correct JavaScript capitalization (#2046)
+ - [readme] Improve eslint config setup instructions for yarn (#2001)
+ - [docs] fix docs for whitespace config (#1914, #1871)
+
 13.1.0 / 2018-08-13
 ==================
  - [new] add eslint v5 support (#1834)